pyproject.toml
[project]
name = "markitdown-rag"
version = "0.1.0"
description = "RAG検証アプリケーション - markitdownを活用したドキュメント変換とRAG機能"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markitdown[all]>=0.1.1",
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
    "langchain-openai>=0.0.2",
    "streamlit>=1.32.0",
    "chromadb>=0.4.22",
    "python-dotenv>=1.0.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.2",
    "mypy>=1.14.1",
    "langchain-chroma",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-cov==6.0.0",
    "ruff==0.11.2",
    "mypy==1.14.1",
]

[tool.ruff]
# Linter と Formatter の設定
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "N", "W", "B", "A", "C4", "UP", "ANN", "RUF"]
ignore = ["ANN101"]  # self の型アノテーションは不要

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src"
